pyproject.toml
[project]
name = "sorting-hat"
version = "0.1.0"
description = "Interview question intent classifier (Coding / System Design / Other) using Qwen + LoRA."
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=2.2",
    "pyarrow>=15",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "litellm>=1.50",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "ipykernel>=6.29",
]
train = [
    "torch>=2.3",
    "transformers>=4.45",
    "peft>=0.13",
    "trl>=0.12",
    "accelerate>=0.33",
    "datasets>=2.18",
    "sentencepiece>=0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sorting_hat"]

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"
